# controller/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Definition of the Frontfacing API of the controller component for\
    \ ecmwf-dataset-crawl. This API is requires no authentication. If required, add\
    \ a proxy for that."
  version: "0.1.0"
  title: "ecmwf-dataset-crawl API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:9000"
basePath: "/api"
tags:
- name: "crawls"
  description: "Crawl Metadata"
- name: "results"
  description: "Crawl Results"
schemes:
- "http"
paths:
  /crawls:
    get:
      tags:
      - "crawls"
      summary: "Get all Crawls"
      operationId: "getCrawls"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CrawlResponse"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "Crawls"
    put:
      tags:
      - "crawls"
      summary: "Add a new Crawl"
      operationId: "addCrawl"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "crawl"
        required: true
        schema:
          $ref: "#/definitions/CrawlRequest"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/CrawlResponse"
        422:
          description: "Malformed Request"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "Crawls"
  /crawls/{crawl-id}:
    get:
      tags:
      - "crawls"
      summary: "Get a Crawl"
      operationId: "getCrawl"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "crawl-id"
        in: "path"
        description: "Crawl ID to operate on"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/CrawlResponse"
        404:
          description: "Crawl not found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "Crawls"
    delete:
      tags:
      - "crawls"
      summary: "Stop a Crawl"
      operationId: "stopCrawl"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "crawl-id"
        in: "path"
        description: "Crawl ID to operate on"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "Crawls"
  /results:
    get:
      tags:
      - "results"
      summary: "Get crawl results with optional filtering"
      description: ""
      operationId: "getResults"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "crawls"
        in: "query"
        description: "Crawl IDs to filter results"
        required: false
        type: "string"
      - name: "query"
        in: "query"
        description: "Elastic Search query term to filter results"
        required: false
        type: "string"
      - name: "maxResults"
        in: "query"
        description: "The amount of results to return"
        required: false
        type: "integer"
        default: 250
      - name: "page"
        in: "query"
        description: "To page through the results"
        required: false
        type: "integer"
        default: 0
      - name: "format"
        in: "query"
        description: "Format in which results are returned"
        required: false
        type: "string"
        enum:
        - "CSV"
      - name: "download"
        in: "query"
        description: "Whether to send an attachment header"
        required: false
        type: "boolean"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Site"
        500:
          description: "Internal Error"
      externalDocs:
        description: "Query parameter from Elastic Search"
        url: "https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl.html"
      x-swagger-router-controller: "Results"
    delete:
      tags:
      - "results"
      summary: "Deletes crawl results based on filter"
      description: ""
      operationId: "deleteResults"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "crawls"
        in: "query"
        description: "Crawl IDs to filter results"
        required: false
        type: "string"
      - name: "query"
        in: "query"
        description: "Elastic Search query term to filter results"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/DeleteResponse"
        500:
          description: "Internal Error"
      externalDocs:
        description: "Query parameter from Elastic Search"
        url: "https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl.html"
      x-swagger-router-controller: "Results"
  /results/counts:
    get:
      tags:
      - "results"
      summary: "Get count of crawl results with optional filtering"
      description: ""
      operationId: "getResultCount"
      produces:
      - "application/json"
      parameters:
      - name: "crawls"
        in: "query"
        description: "Crawl IDs to filter results"
        required: false
        type: "string"
      - name: "query"
        in: "query"
        description: "Elastic Search query term to filter results"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "integer"
        500:
          description: "Internal Error"
      externalDocs:
        description: "Query parameter from Elastic Search"
        url: "https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl.html"
      x-swagger-router-controller: "Results"
definitions:
  CrawlRequest:
    properties:
      languages:
        type: "array"
        items:
          type: "string"
      keywordGroups:
        type: "array"
        items:
          $ref: "#/definitions/CrawlRequest_keywordGroups"
      domainBlacklist:
        type: "array"
        items:
          type: "string"
      domainWhitelist:
        type: "array"
        items:
          type: "string"
    example:
      domainWhitelist:
      - "domainWhitelist"
      - "domainWhitelist"
      domainBlacklist:
      - "domainBlacklist"
      - "domainBlacklist"
      languages:
      - "languages"
      - "languages"
      keywordGroups:
      - keywords:
        - "keywords"
        - "keywords"
        translate: true
      - keywords:
        - "keywords"
        - "keywords"
        translate: true
  CrawlResponse:
    allOf:
    - $ref: "#/definitions/CrawlRequest"
    - properties:
        id:
          type: "string"
          format: "uuid"
        started:
          type: "string"
          format: "date-time"
        completed:
          type: "string"
          format: "date-time"
        seedUrls:
          type: "array"
          items:
            type: "string"
  Site:
    properties:
      crawl:
        type: "string"
        format: "uuid"
      url:
        type: "string"
      scores:
        $ref: "#/definitions/Site_scores"
      metadata:
        $ref: "#/definitions/Site_metadata"
    example:
      metadata:
        contact: "contact"
      scores:
        dataportal: 6.02745618307040320615897144307382404804229736328125
        dataset: 0.80082819046101150206595775671303272247314453125
      url: "https://example.com/foobar"
      crawl: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
  DeleteResponse:
    properties:
      deleted:
        type: "integer"
    example:
      deleted: 0
  CrawlRequest_keywordGroups:
    properties:
      keywords:
        type: "array"
        items:
          type: "string"
          minLength: 1
      translate:
        type: "boolean"
    example:
      keywords:
      - "keywords"
      - "keywords"
      translate: true
  Site_scores:
    properties:
      dataset:
        type: "number"
      dataportal:
        type: "number"
    example:
      dataportal: 6.02745618307040320615897144307382404804229736328125
      dataset: 0.80082819046101150206595775671303272247314453125
  Site_metadata:
    properties:
      contact:
        type: "string"
    example:
      contact: "contact"
externalDocs:
  description: "GitHub"
  url: "https://github.com/52north/ecmwf-dataset-crawl"